chore(data): post_appends->is_allow

+++ database/migrations/2022_12_14_090900_change_post_appends_table.php
@@ -21,7 +21,7 @@ class ChangePostAppendsTable extends Migration
     {
         if (Schema::hasColumn('post_appends', 'is_allow')) {
             Schema::table('post_appends', function (Blueprint $table) {
-                $table->unsignedTinyInteger('is_allow')->default(1)->change();
+                $table->unsignedSmallInteger('is_allow')->default(1)->change();
             });
         }
     }
@@ -35,7 +35,7 @@ class ChangePostAppendsTable extends Migration
     {
         if (Schema::hasColumn('post_appends', 'is_allow')) {
             Schema::table('post_appends', function (Blueprint $table) {
-                $table->unsignedTinyInteger('is_allow')->default(0)->change();
+                $table->unsignedSmallInteger('is_allow')->default(0)->change();
             });
         }
     }
